Now uses metric_term option to disable the use of the metric term adjustments to the advection equation.
This permits the code to revert back to an older style of advection which forces air up more strongly.

Note: This needs to be tested still.  It would also be useful to follow all th jacobian / dx / dz calculations through the code and cleanup any duplication. At various times, the fixed_dz_advection option was used to do something similar.  Its implementation is likely completely out of sync with the current jacobian terms, and should be removed after due analysis and consideration of any related implications.